% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_rde_var}
\alias{load_rde_var}
\title{Title}
\usage{
load_rde_var(useCache = FALSE, loadFcn, cache)
}
\arguments{
\item{useCache}{boolean to force the use of cached data}

\item{loadFcn}{code to load the data from its original source}

\item{cache}{code that acts as a cache for the data}
}
\value{
The data, either loaded using loadFcn, if possible, or from cache if that fails.
}
\description{
Title
}
